% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6pbar.r
\name{formattime}
\alias{formattime}
\title{Time format function}
\usage{
formattime(s)
}
\arguments{
\item{s}{A numerical variable, a time in seconds to convert.}
}
\value{
A character variable, formatted in hh:mm:ss (with two decimals).
}
\description{
Converts time in seconds to time in hh:mm:ss format. If time value passed as
parameter is a float number, only the two first decimals will be kept.
}
\examples{
formattime(32) # outputs '00:00:32.00s'
formattime(165) # outputs '00:02:45.00s'
formattime(3601) # outputs '01:00:01.00s'
formattime(5421.15) # outputs '01:30:21.15s'
formattime(452632.456) # outputs '125:43:52.46s'
}
