% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6pbar.r
\name{\%+\%}
\alias{\%+\%}
\title{String concatenation operator}
\usage{
a \%+\% b
}
\arguments{
\item{a}{A character variable.}

\item{b}{A character variable.}
}
\value{
A character variable, obtained by calling paste0(a,b).
}
\description{
Operator function to make string contatenation easier. This operator does
exactly the same as calling the function \code{paste0(a,b)}
}
\examples{
'Hello' \%+\% ', ' \%+\% 'world !'
}
