% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6pbar.r
\name{shortformatspeed}
\alias{shortformatspeed}
\title{Short speed format function}
\usage{
shortformatspeed(ips)
}
\arguments{
\item{ips}{A numerical variable, a speed in indexes per second to convert.}
}
\value{
A character variable, formatted to short speed
}
\description{
Converts speed in indexes per second to most easily readable two-decimal unit, between
indexes per day (i/d), hour (i/h), minute (i/m) or second (i/s), followed by the corresponding
unit
}
\examples{
shortformatspeed(4564) # outputs '4564 i/s'
shortformatspeed(45) # outputs '45 i/s'
shortformatspeed(48/60) # outputs '48 i/m'
shortformatspeed(3/60) # outputs '3 i/m'
shortformatspeed(45/3600) # outputs '45 i/h'
shortformatspeed(0.5/3600) # outputs '12 i/d'
}
