% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6pbar.r
\name{shortformattime}
\alias{shortformattime}
\title{Short time format function}
\usage{
shortformattime(s)
}
\arguments{
\item{s}{A numerical variable, a time in seconds to convert.}
}
\value{
A character variable, formatted to short time
}
\description{
Converts time in seconds to most easily readable non-decimal unit, between
days, hours, minutes or seconds, followed by the corresponding unit
}
\examples{
shortformattime(32) # outputs '32s'
shortformattime(165) # outputs '3m'
shortformattime(3601) # outputs '1h'
shortformattime(5421.15) # outputs '2h'
shortformattime(452632.456) # outputs '5d'
}
